requirements.txt
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 # 
numba
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
scikit-image==0.19.3
facexlib==0.3.0
gradio==4.25.0
cython
gfpgan
av
safetensors
python-dotenv
rq
rq_dashboard_fast
azure-storage-blob
-e git+https://github.com/smetak13/BasicSR.git#egg=basicsr